wrapper implementation for chain registry
- ChainProviderProtocol conformance implementation
- RuntimeCodingServiceProviderProtocol conformance implementation
