--- novawallet/Common/Services/ChainProvider/ChainRegistry+ChainProvider.swift
@@ -1,12 +1,0 @@
-import Foundation
-import Operation_iOS
-
-extension ChainRegistry: ChainProviderProtocol {
-    func createChainWrapper(for chainId: ChainModel.Id) -> CompoundOperationWrapper<ChainModel> {
-        do {
-            return .createWithResult(try getChainOrError(for: chainId))
-        } catch {
-            return .createWithError(error)
-        }
-    }
-}
